+++ Sources/Storage/MultipartFormData.swift
@@ -1,4 +1,4 @@
-// MutlipartFormData extracted from [Alamofire](https://github.com/Alamofire/Alamofire/blob/master/Source/Features/MultipartFormData.swift) for using as standalone.
+// MultipartFormData extracted from [Alamofire](https://github.com/Alamofire/Alamofire/blob/master/Source/Features/MultipartFormData.swift) for using as standalone.
 
 //
 //  MultipartFormData.swift