+++ Tests/RealtimeTests/RealtimeTests.swift
@@ -175,6 +175,7 @@ import XCTest
                 "config" : {
                   "broadcast" : {
                     "ack" : false,
+                    "replication_ready" : false,
                     "self" : false
                   },
                   "postgres_changes" : [
@@ -267,6 +268,7 @@ import XCTest
               "config" : {
                 "broadcast" : {
                   "ack" : false,
+                  "replication_ready" : false,
                   "self" : false
                 },
                 "postgres_changes" : [
@@ -291,6 +293,7 @@ import XCTest
               "config" : {
                 "broadcast" : {
                   "ack" : false,
+                  "replication_ready" : false,
                   "self" : false
                 },
                 "postgres_changes" : [
@@ -674,7 +677,7 @@ import XCTest
         	--header "Authorization: Bearer custom.access.token" \
         	--header "Content-Type: application/json" \
         	--header "apiKey: publishable.api.key" \
-        	--data "{\"messages\":[{\"event\":\"test\",\"payload\":{\"value\":42},\"private\":false,\"topic\":\"realtime:public:messages\"}]}" \
+        	--data "{\"messages\":[{\"event\":\"test\",\"payload\":{\"value\":42},\"private\":false,\"topic\":\"public:messages\"}]}" \
         	"http://localhost:54321/realtime/v1/api/broadcast"
         """#
       }